Source code for the T41-EP Project done by Jack Purdum, W8TEE, and Al Peter, AC8GY.
NOTE: The source code is not really done yet. With about 12,000 lines of C code, I'm not sure it will ever be "done". I fully expect this code to morph into something really great over time. I am still not happy with the CW xmtr portion of the code and am experimenting with both polling and interrupts. I will post what I come up with when it's done, and maybe some of you will find solutions that are even better. If so, please let us know.